5.14. Модуляция (домашнее задание)
