+++ js/setup.js
@@ -1,0 +1,5 @@
+const PHOTOCARDS_COUNT = 25; // Количество фотографий
+const MAX_COMMENTS_COUNT = 30; // Максимальное число комментариев к фото
+const MAX_LIKES_COUNT = 30; // Максимальное число лайков к фото
+
+export { PHOTOCARDS_COUNT, MAX_COMMENTS_COUNT, MAX_LIKES_COUNT };
